Add 'divisibleSumPairs' function in 'Result.java'
The function calculates the number of pairs in an array where the sum is evenly divisible by a specified number. The 'divisibleSumPairs' function accepts a list of integers, and two separate integers representing the size of the list and the divisor, respectively. The implemented solution uses the concept of modular arithmetic to achieve this.